requirements.txt
yt-dlp
youtube-transcript-api
pdfplumber
google-auth-oauthlib
google-api-python-client
